#1067 Fix Double -> BigDecimal calling createInventoryItem service
git-svn-id: svn://svn.opentaps.org/opentaps_all/branches/ofbiz-0904-1.4.0-preview1@13025 d3523486-f5fe-0310-a41a-a15a4e76f3c7

+++ hot-deploy/warehouse/src/org/opentaps/warehouse/domain/inventory/InventoryService.java
@@ -237,7 +237,7 @@ public class InventoryService extends Service implements InventoryServiceInterfa
                 input.remove("oldAvailableToPromiseTotal");
                 input.remove("oldQuantityOnHandTotal");
                 // convert BigDecimal to back their expected values
-                input.put("unitCost", sourceInventoryItem.getUnitCost().doubleValue());
+                input.put("unitCost", sourceInventoryItem.getUnitCost());
                 Map<String, Object> tmpResult = runSync("createInventoryItem", input);
                 String newInventoryItemId = (String) tmpResult.get("inventoryItemId");
